Add test_cancellation.py
